AssitantBuilder: Actions components
